requirements.txt

streamlit==1.42.2
torch==2.6.0
pandas==2.2.3
transformers==4.49.0
openai==1.9.0
gspread==6.2.0
oauth2client==4.1.3
python-dotenv==1.0.1
